BOOKING-201: Moved tabs and filters out of app.js
