+++ assets/src/create-booking.js
@@ -1,5 +1,4 @@
 import React, { useEffect, useState } from "react";
-import Select, { createFilter } from "react-select";
 import dayjs from "dayjs";
 import "dayjs/locale/da";
 import { useSearchParams } from "react-router-dom";
@@ -15,10 +14,11 @@ import MapWrapper from "./components/map-wrapper";
 import MainNavigation from "./components/main-navigation";
 import Api from "./util/api";
 import UrlValidator from "./util/url-validator";
-import { capacityOptions } from "./util/filter-utils";
-import { hasOwnProperty, filterAllResources, getFacilityOptions } from "./util/helpers";
+import { hasOwnProperty, filterAllResources } from "./util/helpers";
 import { setAriaLabelFilters } from "./util/dom-manipulation-utils";
 import "react-toastify/dist/ReactToastify.css";
+import CreateBookingFilters from "./create-booking-filters";
+import CreateBookingTabs from "./components/create-booking-tabs";
 
 dayjs.locale("da");
 
@@ -30,36 +30,30 @@ dayjs.locale("da");
  * @returns {JSX.Element} Component.
  */
 function CreateBooking({ config }) {
-  // App configuration and behavior.
-  const [displayState, setDisplayState] = useState("maximized"); // The app display mode to be used.
-  const [urlResource, setUrlResource] = useState(null); // A resource fetched from API if validUrlParams are set.
-  const [validUrlParams, setValidUrlParams] = useState(null); // Validated url params through url-validator.js.
-  const [urlParams] = useSearchParams(); // Url parameters when the app is loaded.
-  const [bookingView, setBookingView] = useState("calendar");
-  // Options for filters.
-  const [locationOptions, setLocationOptions] = useState([]);
-  const [resourcesOptions, setResourcesOptions] = useState([]);
-  // User selections in the filters.
-  const [date, setDate] = useState(new Date()); // Date filter selected in calendar header component.
-  const [filterParams, setFilterParams] = useState({}); // An object containing structured information about current filtering.
+  const [urlParams] = useSearchParams();
+  // Booking data.
+  const [date, setDate] = useState(new Date());
+  const [calendarSelection, setCalendarSelection] = useState({});
+  const [authorFields, setAuthorFields] = useState({ subject: "", email: "" });
+  // Filter parameters, selected in CreateBookingFilters. An object containing
+  // structured information about current filtering.
+  const [filterParams, setFilterParams] = useState({});
   const [locationFilter, setLocationFilter] = useState([]);
   const [resourceFilter, setResourceFilter] = useState([]);
-  const [capacityFilter, setCapacityFilter] = useState([]);
-  const [facilityFilter, setFacilityFilter] = useState([]);
-  // App display for calendar, list and map.
-  const [events, setEvents] = useState([]); // Events related to the displayed resources (free/busy).
-  // Resources need to be false until we set it the first time, because [] equals no results and false triggers placeholder resources.
-  // TODO: Handle this in another way so the propType does not throw a warning.
-  const [resources, setResources] = useState(null); // The result after filtering resources
-  const [showResourceDetails, setShowResourceDetails] = useState(null); // ID of the displayed resource.
-  // eslint-disable-next-line no-unused-vars
-  const [allResources, setAllResources] = useState([]);
-  // App output. - Data to be pushed to API or used as parameters for redirect.
-  const [authorFields, setAuthorFields] = useState({ subject: "", email: "" }); // Additional fields for author information.
-  const [calendarSelection, setCalendarSelection] = useState({}); // The selection of a time span in calendar.
+  // App configuration and behavior.
+  const [displayState, setDisplayState] = useState("maximized");
+  const [urlResource, setUrlResource] = useState(null);
+  const [validUrlParams, setValidUrlParams] = useState(null);
+  const [activeTab, setActiveTab] = useState("calendar");
   const [userHasInteracted, setUserHasInteracted] = useState(false);
   const [isLoading, setIsLoading] = useState(false);
+  const [showResourceDetails, setShowResourceDetails] = useState(null);
+  // Loaded data.
+  const [events, setEvents] = useState([]);
+  const [resources, setResources] = useState(null);
+  const [allResources, setAllResources] = useState([]);
 
+  // TODO: Describe.
   setAriaLabelFilters();
 
   // Get configuration.
@@ -71,38 +65,8 @@ function CreateBooking({ config }) {
     }
   }, []);
 
-  // Loop allresources and set filter options
-  useEffect(() => {
-    if (resourcesOptions.length === allResources.length) {
-      return;
-    }
-
-    const locations = [
-      ...new Set(allResources.filter((resource) => resource.location !== "").map((resource) => resource.location)),
-    ];
-
-    setLocationOptions(
-      locations
-        .map((value) => {
-          return {
-            value,
-            label: value,
-          };
-        })
-        .sort()
-    );
-
-    setResourcesOptions(
-      allResources.map((value) => {
-        return {
-          value: value.resourceMail,
-          label: value.resourceName,
-        };
-      })
-    );
-  }, [allResources]);
-
-  // Effects to run when config is loaded. This should only happen once at app initialisation.
+  // Effects to run when config is loaded. This should only happen once at app
+  // initialisation.
   useEffect(() => {
     if (config) {
       Api.fetchAllResources(config.api_endpoint)
@@ -125,7 +89,8 @@ function CreateBooking({ config }) {
     }
   }, [config]);
 
-  // Effects to run when urlResource is set. This should only happen once in extension of app initialisation.
+  // Effects to run when urlResource is set. This should only happen once in
+  // extension of app initialisation.
   useEffect(() => {
     // Set location filter.
     if (urlResource && hasOwnProperty(urlResource, "location")) {
@@ -160,7 +125,7 @@ function CreateBooking({ config }) {
     }
   }, [urlResource]);
 
-  // Set resources from filterParams.
+  // Find resources that match filterParams.
   useEffect(() => {
     if (config) {
       const urlSearchParams = new URLSearchParams();
@@ -193,74 +158,6 @@ function CreateBooking({ config }) {
     }
   }, [filterParams]);
 
-  // Set location filter and resource dropdown options.
-  useEffect(() => {
-    const locationValues = locationFilter.map(({ value }) => value);
-
-    setFilterParams({ ...filterParams, ...{ "location[]": locationValues } });
-  }, [locationFilter]);
-
-  // Set resource filter.
-  useEffect(() => {
-    const resourceValues = resourceFilter.map(({ value }) => value);
-
-    setFilterParams({
-      ...filterParams,
-      ...{ "resourceMail[]": resourceValues },
-    });
-  }, [resourceFilter]);
-
-  // Set capacity filter.
-  useEffect(() => {
-    const newFilterParams = { ...filterParams };
-    const capacityType = capacityFilter.type ?? null;
-    const capacityValue = capacityFilter.value ?? 0;
-
-    // Delete opposite entry to prevent both capacity[between] and capacity[gt] being set, causing a dead end.
-    delete newFilterParams["capacity[between]"];
-
-    delete newFilterParams["capacity[gt]"];
-
-    // Set capacity filter according to capacityType.
-    let capacity;
-
-    switch (capacityType) {
-      case "between":
-        capacity = { "capacity[between]": capacityValue };
-
-        break;
-      case "gt":
-        capacity = { "capacity[gt]": capacityValue };
-
-        break;
-      default:
-        break;
-    }
-
-    setFilterParams({ ...newFilterParams, ...capacity });
-  }, [capacityFilter]);
-
-  // Set facility filter.
-  useEffect(() => {
-    const filterParamsObj = { ...filterParams };
-
-    delete filterParamsObj.monitorEquipment;
-
-    delete filterParamsObj.wheelchairAccessible;
-
-    delete filterParamsObj.videoConferenceEquipment;
-
-    delete filterParamsObj.catering;
-
-    const facilitiesObj = {};
-
-    facilityFilter.forEach((value) => {
-      facilitiesObj[value.value] = "true";
-    });
-
-    setFilterParams({ ...filterParamsObj, ...facilitiesObj });
-  }, [facilityFilter]);
-
   // Get events for the given resources.
   useEffect(() => {
     if (config && resources?.length > 0 && date !== null) {
@@ -292,229 +189,121 @@ function CreateBooking({ config }) {
     }
   }, [calendarSelection, authorFields]);
 
-  const viewSwapHandler = (event) => {
-    const view = event.target.getAttribute("data-view");
-
-    setBookingView(view);
-  };
-
   return (
-    <div>
+    config && (
       <div className="App">
         <div className="container-fluid">
-          {config && <MainNavigation config={config} />}
+          <MainNavigation config={config} />
           <div className="app-wrapper">
-            {config && (
-              <div>
-                {config && config && displayState === "maximized" && (
-                  <div className="app-content">
-                    <div className={`row filters-wrapper ${showResourceDetails !== null ? "disable-filters" : ""}`}>
-                      <div className="col-md-3 col-xs-12 small-padding">
-                        <label htmlFor="location-filter">
-                          Filtrér på lokationer
-                          {/* Dropdown with locations */}
-                          <Select
-                            styles={{}}
-                            id="location-filter"
-                            className="filter"
-                            defaultValue={locationFilter}
-                            value={locationFilter}
-                            placeholder="lokationer..."
-                            placeholderClassName="dropdown-placeholder"
-                            closeMenuOnSelect={false}
-                            options={locationOptions}
-                            onChange={(selectedLocations) => {
-                              setLocationFilter(selectedLocations);
-                            }}
-                            isLoading={Object.values(locationOptions).length === 0}
-                            loadingMessage={() => "Henter lokationer.."}
-                            filterOption={createFilter({ ignoreAccents: false })} // Improved performance with large datasets
-                            isMulti
-                          />
-                        </label>
-                      </div>
-                      <div className="col-md-3 col-xs-12 small-padding">
-                        <label htmlFor="resource-filter">
-                          Filtrér på lokaler/resource
-                          {/* Dropdown with resources */}
-                          <Select
-                            styles={{}}
-                            id="resource-filter"
-                            className="filter"
-                            defaultValue={resourceFilter}
-                            placeholder="ressourcer..."
-                            placeholderClassName="dropdown-placeholder"
-                            closeMenuOnSelect={false}
-                            options={resourcesOptions}
-                            onChange={(selectedResources) => {
-                              setResourceFilter(selectedResources);
-                            }}
-                            isLoading={Object.values(resourcesOptions).length === 0}
-                            loadingMessage={() => "Henter ressourcer.."}
-                            filterOption={createFilter({ ignoreAccents: false })} // Improved performance with large datasets
-                            isMulti
-                          />
-                        </label>
-                      </div>
-                      <div className="col-md-3 col-xs-12 small-padding">
-                        <label htmlFor="facility-filter">
-                          Filtrér på faciliteter
-                          {/* Dropdown with facilities */}
-                          <Select
-                            styles={{}}
-                            id="facility-filter"
-                            className="filter"
-                            defaultValue={facilityFilter}
-                            placeholder="Facilitieter..."
-                            placeholderClassName="dropdown-placeholder"
-                            closeMenuOnSelect={false}
-                            options={getFacilityOptions()}
-                            onChange={(selectedFacilities) => {
-                              setFacilityFilter(selectedFacilities);
-                            }}
-                            isMulti
-                          />
-                        </label>
-                      </div>
-                      <div className="col-md-3 col-xs-12 small-padding">
-                        <label htmlFor="capacity-filter">
-                          Filtrér på kapacitet
-                          {/* Dropdown with capacity */}
-                          <Select
-                            styles={{}}
-                            id="capacity-filter"
-                            className="filter"
-                            defaultValue={{ value: "0", label: "Alle", type: "gt" }}
-                            placeholder="Siddepladser..."
-                            placeholderClassName="dropdown-placeholder"
-                            closeMenuOnSelect
-                            options={capacityOptions}
-                            onChange={(selectedCapacity) => {
-                              setCapacityFilter(selectedCapacity);
-                            }}
-                          />
-                        </label>
-                      </div>
-                    </div>
+            <div>
+              {displayState === "maximized" && (
+                <div className="app-content">
+                  <CreateBookingFilters
+                    filterParams={filterParams}
+                    setFilterParams={setFilterParams}
+                    allResources={allResources}
+                    disabled={showResourceDetails ?? false}
+                    locationFilter={locationFilter}
+                    setLocationFilter={setLocationFilter}
+                    resourceFilter={resourceFilter}
+                    setResourceFilter={setResourceFilter}
+                  />
 
-                    {/* Add info box */}
-                    <div className="row info-box-wrapper">
-                      {config.info_box_color && config.info_box_header && config.info_box_content && (
-                        <InfoBox config={config} />
-                      )}
-                    </div>
-
-                    {/* Add viewswapper */}
-                    <div className="row viewswapper-wrapper">
-                      <div className="viewswapper-container">
-                        <button
-                          type="button"
-                          onClick={viewSwapHandler}
-                          data-view="calendar"
-                          className={bookingView === "calendar" ? "active booking-btn" : "booking-btn"}
-                        >
-                          Kalender
-                        </button>
-                        <button
-                          type="button"
-                          onClick={viewSwapHandler}
-                          data-view="list"
-                          className={bookingView === "list" ? "active booking-btn" : "booking-btn"}
-                        >
-                          Liste
-                        </button>
-                        <button
-                          type="button"
-                          onClick={viewSwapHandler}
-                          data-view="map"
-                          className={bookingView === "map" ? "active booking-btn" : "booking-btn"}
-                        >
-                          Kort
-                        </button>
-                      </div>
-                    </div>
-
-                    {bookingView === "map" && (
-                      <div className="row no-gutter main-container map">
-                        <MapWrapper
-                          allResources={allResources}
-                          config={config}
-                          setLocationFilter={setLocationFilter}
-                          setBookingView={setBookingView}
-                        />
-                      </div>
-                    )}
-                    {bookingView === "list" && (
-                      <div
-                        className={`row no-gutter main-container list ${
-                          showResourceDetails !== null ? "resourceview-visible" : ""
-                        }`}
-                      >
-                        <ListContainer
-                          resources={resources}
-                          setShowResourceDetails={setShowResourceDetails}
-                          isLoading={isLoading}
-                        />
-                        <ResourceView
-                          showResourceDetails={showResourceDetails}
-                          setShowResourceDetails={setShowResourceDetails}
-                        />
-                      </div>
-                    )}
-                    {bookingView === "calendar" && (
-                      // {/* Display calendar for selections */}
-                      <div
-                        className={`row no-gutter main-container calendar ${
-                          showResourceDetails !== null ? "resourceview-visible" : ""
-                        }`}
-                      >
-                        <Calendar
-                          resources={resources}
-                          events={events}
-                          date={date}
-                          setDate={setDate}
-                          calendarSelection={calendarSelection}
-                          setCalendarSelection={setCalendarSelection}
-                          config={config}
-                          setShowResourceDetails={setShowResourceDetails}
-                          urlResource={urlResource}
-                          setDisplayState={setDisplayState}
-                          showResourceDetails={showResourceDetails}
-                          userHasInteracted={userHasInteracted}
-                          isLoading={isLoading}
-                          setIsLoading={setIsLoading}
-                        />
-                        {/* TODO: Only show if resource view is requested */}
-                        <ResourceView
-                          showResourceDetails={showResourceDetails}
-                          setShowResourceDetails={setShowResourceDetails}
-                        />
-                      </div>
+                  {/* Add info box */}
+                  <div className="row info-box-wrapper">
+                    {config.info_box_color && config.info_box_header && config.info_box_content && (
+                      <InfoBox config={config} />
                     )}
                   </div>
-                )}
 
-                {config && urlResource && displayState === "minimized" && calendarSelection && (
-                  <div className="row">
-                    <MinimizedDisplay
-                      setDisplayState={setDisplayState}
-                      urlResource={urlResource}
-                      calendarSelection={calendarSelection}
-                    />
-                  </div>
-                )}
-              </div>
-            )}
+                  <CreateBookingTabs activeTab={activeTab} setActiveTab={setActiveTab} />
+
+                  {/* Map view */}
+                  {activeTab === "map" && (
+                    <div className="row no-gutter main-container map">
+                      <MapWrapper
+                        allResources={allResources}
+                        config={config}
+                        setLocationFilter={setLocationFilter}
+                        setBookingView={setActiveTab}
+                      />
+                    </div>
+                  )}
+
+                  {/* List view */}
+                  {activeTab === "list" && (
+                    <div
+                      className={`row no-gutter main-container list ${
+                        showResourceDetails !== null ? "resourceview-visible" : ""
+                      }`}
+                    >
+                      <ListContainer
+                        resources={resources}
+                        setShowResourceDetails={setShowResourceDetails}
+                        isLoading={isLoading}
+                      />
+                      <ResourceView
+                        showResourceDetails={showResourceDetails}
+                        setShowResourceDetails={setShowResourceDetails}
+                      />
+                    </div>
+                  )}
+
+                  {/* Calendar view */}
+                  {activeTab === "calendar" && (
+                    // {/* Display calendar for selections */}
+                    <div
+                      className={`row no-gutter main-container calendar ${
+                        showResourceDetails !== null ? "resourceview-visible" : ""
+                      }`}
+                    >
+                      <Calendar
+                        resources={resources}
+                        events={events}
+                        date={date}
+                        setDate={setDate}
+                        calendarSelection={calendarSelection}
+                        setCalendarSelection={setCalendarSelection}
+                        config={config}
+                        setShowResourceDetails={setShowResourceDetails}
+                        urlResource={urlResource}
+                        setDisplayState={setDisplayState}
+                        showResourceDetails={showResourceDetails}
+                        userHasInteracted={userHasInteracted}
+                        isLoading={isLoading}
+                        setIsLoading={setIsLoading}
+                      />
+
+                      <ResourceView
+                        showResourceDetails={showResourceDetails}
+                        setShowResourceDetails={setShowResourceDetails}
+                      />
+                    </div>
+                  )}
+                </div>
+              )}
+
+              {urlResource && displayState === "minimized" && calendarSelection && (
+                <div className="row">
+                  <MinimizedDisplay
+                    setDisplayState={setDisplayState}
+                    urlResource={urlResource}
+                    calendarSelection={calendarSelection}
+                  />
+                </div>
+              )}
+            </div>
           </div>
         </div>
+
+        {/* Display author fields if user is logged in */}
+        {!config?.step_one && (
+          <div className="row no-gutter">
+            {authorFields && <AuthorFields authorFields={authorFields} setAuthorFields={setAuthorFields} />}
+          </div>
+        )}
       </div>
-      {/* Display author fields */}
-      {config && !config.step_one && (
-        <div className="row no-gutter">
-          {authorFields && <AuthorFields authorFields={authorFields} setAuthorFields={setAuthorFields} />}
-        </div>
-      )}
-    </div>
+    )
   );
 }
